requirements.txt
Django>=3.2.4,<3.3
djangorestframework >=3.12.4,<13.13
mysqlclient>=2.0.3,<2.1
pytz~=2023.3